+++ core/data/src/commonMain/kotlin/com/whatever/caramel/core/data/mapper/AuthMapper.kt
@@ -3,8 +3,8 @@ package com.whatever.caramel.core.data.mapper
 import com.whatever.caramel.core.domain.vo.auth.AuthToken
 import com.whatever.caramel.core.domain.vo.auth.UserAuth
 import com.whatever.caramel.core.domain.vo.user.UserStatus
-import com.whatever.caramel.core.remote.dto.auth.ServiceToken
-import com.whatever.caramel.core.remote.dto.auth.SignInResponse
+import com.whatever.caramel.core.remote.dto.auth.ServiceTokenDto
+import com.whatever.caramel.core.remote.dto.auth.response.SignInResponse
 import com.whatever.caramel.core.util.DateParser.toMillisecond
 
 fun SignInResponse.toUserAuth(): UserAuth {
@@ -17,7 +17,7 @@ fun SignInResponse.toUserAuth(): UserAuth {
     )
 }
 
-fun ServiceToken.toAuthToken() = AuthToken(
+fun ServiceTokenDto.toAuthToken() = AuthToken(
     accessToken = this.accessToken,
     refreshToken = this.refreshToken
 )
