+++ core/data/src/commonMain/kotlin/com/whatever/caramel/core/data/mapper/CoupleMapper.kt
@@ -9,10 +9,10 @@ import com.whatever.caramel.core.domain.vo.couple.CoupleRelationship
 import com.whatever.caramel.core.domain.vo.couple.CoupleStatus
 import com.whatever.caramel.core.domain.vo.user.Gender
 import com.whatever.caramel.core.domain.vo.user.UserProfile
+import com.whatever.caramel.core.remote.dto.couple.CoupleUserInfoDto
 import com.whatever.caramel.core.remote.dto.couple.response.CoupleBasicResponse
 import com.whatever.caramel.core.remote.dto.couple.response.CoupleDetailResponse
 import com.whatever.caramel.core.remote.dto.couple.response.CoupleInvitationCodeResponse
-import com.whatever.caramel.core.remote.dto.couple.response.CoupleUserInfoDto
 import com.whatever.caramel.core.util.DateParser.toMillisecond
 import kotlinx.datetime.Instant
