+++ core/data/src/commonMain/kotlin/com/whatever/caramel/core/data/repository/UserRepositoryImpl.kt
@@ -9,8 +9,8 @@ import com.whatever.caramel.core.domain.repository.UserRepository
 import com.whatever.caramel.core.domain.vo.user.Gender
 import com.whatever.caramel.core.domain.vo.user.UserStatus
 import com.whatever.caramel.core.remote.datasource.RemoteUserDataSource
-import com.whatever.caramel.core.remote.dto.user.EditUserProfileRequest
-import com.whatever.caramel.core.remote.dto.user.UserProfileRequest
+import com.whatever.caramel.core.remote.dto.user.request.EditUserProfileRequest
+import com.whatever.caramel.core.remote.dto.user.request.UserProfileRequest
 
 class UserRepositoryImpl(
     private val userRemoteDataSource: RemoteUserDataSource,
