+++ core/remote/src/commonMain/kotlin/com/whatever/caramel/core/remote/datasource/RemoteUserDataSource.kt
@@ -1,9 +1,9 @@
 package com.whatever.caramel.core.remote.datasource
 
-import com.whatever.caramel.core.remote.dto.user.EditUserProfileRequest
-import com.whatever.caramel.core.remote.dto.user.EditUserProfileResponse
-import com.whatever.caramel.core.remote.dto.user.UserProfileRequest
-import com.whatever.caramel.core.remote.dto.user.UserProfileResponse
+import com.whatever.caramel.core.remote.dto.user.request.EditUserProfileRequest
+import com.whatever.caramel.core.remote.dto.user.response.EditUserProfileResponse
+import com.whatever.caramel.core.remote.dto.user.request.UserProfileRequest
+import com.whatever.caramel.core.remote.dto.user.response.UserProfileResponse
 
 interface RemoteUserDataSource {
     suspend fun createUserProfile(request: UserProfileRequest): UserProfileResponse
