+++ core/remote/src/commonMain/kotlin/com/whatever/caramel/core/remote/datasource/RemoteUserDataSourceImpl.kt
@@ -1,9 +1,9 @@
 package com.whatever.caramel.core.remote.datasource
 
-import com.whatever.caramel.core.remote.dto.user.EditUserProfileRequest
-import com.whatever.caramel.core.remote.dto.user.EditUserProfileResponse
-import com.whatever.caramel.core.remote.dto.user.UserProfileRequest
-import com.whatever.caramel.core.remote.dto.user.UserProfileResponse
+import com.whatever.caramel.core.remote.dto.user.request.EditUserProfileRequest
+import com.whatever.caramel.core.remote.dto.user.response.EditUserProfileResponse
+import com.whatever.caramel.core.remote.dto.user.request.UserProfileRequest
+import com.whatever.caramel.core.remote.dto.user.response.UserProfileResponse
 import com.whatever.caramel.core.remote.network.util.getBody
 import io.ktor.client.HttpClient
 import io.ktor.client.request.post
