+++ core/remote/src/commonMain/kotlin/com/whatever/caramel/core/remote/dto/auth/LoginPlatformDto.kt
@@ -4,13 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SignInRequest(
-    @SerialName("loginPlatform") val loginPlatform: LoginPlatform,
-    @SerialName("idToken") val idToken: String,
-)
-
-@Serializable
-enum class LoginPlatform {
+enum class LoginPlatformDto {
     @SerialName("KAKAO") KAKAO,
     @SerialName("APPLE") APPLE,
     ;
