+++ core/remote/src/commonMain/kotlin/com/whatever/caramel/core/remote/dto/auth/ServiceTokenDto.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ServiceToken(
+data class ServiceTokenDto(
     @SerialName("accessToken") val accessToken: String,
     @SerialName("refreshToken") val refreshToken: String,
 )
