+++ core/remote/src/commonMain/kotlin/com/whatever/caramel/core/remote/dto/couple/response/CoupleInfoResponse.kt
@@ -1,6 +1,7 @@
 package com.whatever.caramel.core.remote.dto.couple.response
 
-import com.whatever.caramel.core.remote.dto.user.response.GenderDto
+import com.whatever.caramel.core.remote.dto.couple.CoupleStatusDto
+import com.whatever.caramel.core.remote.dto.couple.CoupleUserInfoDto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -21,18 +22,3 @@ data class CoupleBasicResponse(
     @SerialName("sharedMessage") val sharedMessage: String?,
     @SerialName("status") val status: CoupleStatusDto,
 )
-
-@Serializable
-data class CoupleUserInfoDto(
-    @SerialName("id") val id: Long,
-    @SerialName("nickname") val nickname: String,
-    @SerialName("birthDate") val birthDate: String,
-    @SerialName("gender") val gender: GenderDto,
-)
-
-@Serializable
-enum class CoupleStatusDto {
-    @SerialName("ACTIVE") ACTIVE,
-    @SerialName("INACTIVE") INACTIVE,
-    ;
-}
