+++ core/remote/src/commonMain/kotlin/com/whatever/caramel/core/remote/dto/user/request/EditUserProfileRequest.kt
@@ -1,4 +1,4 @@
-package com.whatever.caramel.core.remote.dto.user
+package com.whatever.caramel.core.remote.dto.user.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
