+++ core/remote/src/commonMain/kotlin/com/whatever/caramel/core/remote/dto/user/response/EditUserProfileResponse.kt
@@ -1,4 +1,4 @@
-package com.whatever.caramel.core.remote.dto.user
+package com.whatever.caramel.core.remote.dto.user.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
